al-0.7     * Adaptivity logic is reconstructed to deploy my advisor's midnight hack (see alx.h and al.c.)     * A new benchmark is integrated, test2.c (hash table) which is based on Reimer's implementation.

+++ test3.c
@@ -81,7 +81,7 @@ static int iter = 100000;
 void*
 task(void* arg)
 {
-  unsigned short id = (unsigned short)(unsigned int)arg;
+  unsigned short id = (unsigned short)arg;
   long n = iter;
   long t;
   unsigned short xseed[3] = {id,id,id};
